Set unit tests with pytest
