+++ QuantumPioneer/__init__.py
@@ -1,7 +1,4 @@
 """Automated Workflow for High-Throughput Quantum Chemical Kinetics"""
 
 # Add imports here
-from .main import *
-
-
 from ._version import __version__
