+++ QuantumPioneer/tests/test_QuantumPioneer.py
@@ -5,9 +5,9 @@ Unit and regression test for the QuantumPioneer package.
 # Import package, test suite, and other packages as needed
 import sys
 
-import pytest
+# import pytest
 
-import QuantumPioneer
+# import QuantumPioneer
 
 
 def test_QuantumPioneer_imported():
