+++ pyproject.toml
@@ -93,3 +93,8 @@ ignore = "E203,W503"
 
 [tool.pylint]
 rcfile = ".pylintrc"
+
+[tool.pytest.ini_options]
+minversion = "7.0"
+addopts = "-v -ra --doctest-modules --cov=QuantumPioneer --cov-report=term --cov-report=html"
+testpaths = ["QuantumPioneer"]
